Attempted another solution, I thought I was on a bit of a different path - but this turned into another greedy solution.
Committing this to show a bit of my thought process.

Added some failing tests to highlight the flaws
with this approach.
The easiest one to understand is when you have
the coins 1, 6 and 10 and the ammount we are
looking for is 12. 6+6 is the best solution,
but this currently implmentation chooses 10+1+1