Implmented top down solution
The idea here is to take the target value and subtract coins,
until we reach 0 or run out of candidates.

I think this idea is sound but the implmentation could do
with some tidy up. I've scattered it with HashSets to try
and reduce the cost of checking if this number has been seen
before.